python/requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
cryptography>=41.0.0
pytest>=7.4.0

# Optional — install for sr25519 key support:
# py-sr25519-bindings>=0.2.0

# Optional — install for secp256k1 key support:
# eth-keys>=0.5.0
